backend/pyproject.toml
[project]
name = "cv-gen"
version = "0.1.0"
description = "Generador de CVs en PDF desde Markdown"
requires-python = ">=3.10"
dependencies = [
    "click>=8.1",
    "fastapi>=0.115",
    "jinja2>=3.1",
    "markdown>=3.5",
    "python-docx>=1.1",
    "python-frontmatter>=1.0",
    "python-multipart>=0.0.9",
    "uvicorn[standard]>=0.30",
    "weasyprint>=60.0",
]

[project.optional-dependencies]
ai-google = ["google-genai>=1.0"]
ai-openai = ["openai>=1.0"]
ai-anthropic = ["anthropic>=0.30"]
ai = ["cv-gen[ai-google,ai-openai,ai-anthropic]"]

[project.scripts]
cv-gen = "cv_gen.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/cv_gen"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[dependency-groups]
dev = [
    "httpx>=0.27",
    "pytest>=7.0",
]
